Refactor model loading by introducing ModelLoader; streamline model retrieval in BenchmarkCommand and RunCommand
