+++ Sources/LocalLLMCLI/RunCommand.swift
@@ -54,14 +54,16 @@ struct RunCommand: AsyncParsableCommand {
     func run() async throws {
         let backend = Backend(rawValue: backend) ?? .llama
         log("Loading model from: \(model) with backend: \(backend.rawValue)")
+        
+        let modelLoader = ModelLoader(verbose: verbose)
 
         // Initialize client
         let client: any LLMClient
         switch backend {
         case .llama:
             client = try await LocalLLMClient.llama(
-                url: getModel(for: model, backend: backend),
-                mmprojURL: mmproj.asyncMap { try await getModel(for: $0, backend: backend) },
+                url: modelLoader.getModel(for: model, backend: backend),
+                mmprojURL: mmproj.asyncMap { try await modelLoader.getModel(for: $0, backend: backend) },
                 parameter: .init(
                     temperature: temperature,
                     topK: topK,
@@ -72,7 +74,7 @@ struct RunCommand: AsyncParsableCommand {
         case .mlx:
 #if canImport(LocalLLMClientMLX)
             client = try await LocalLLMClient.mlx(
-                url: getModel(for: model, backend: backend),
+                url: modelLoader.getModel(for: model, backend: backend),
                 parameter: .init(
                     temperature: temperature,
                     topP: topP,
@@ -117,45 +119,6 @@ struct RunCommand: AsyncParsableCommand {
 
         log("\n---")
         log("Generation complete.")
-    }
-
-    private func getModel(for model: String, backend: Backend) async throws -> URL {
-        return if model.hasPrefix("/") {
-            URL(filePath: model)
-        } else if model.hasPrefix("https://"), let url = URL(string: model) {
-            try await downloadModel(from: url, backend: backend)
-        } else if model.isEmpty {
-            throw LocalLLMCommandError.invalidModel("Error: Missing expected argument '--model <model>'")
-        } else {
-            throw LocalLLMCommandError.invalidModel(model)
-        }
-    }
-
-    private func downloadModel(from url: URL, backend: Backend) async throws -> URL {
-        #if canImport(LocalLLMClientUtility)
-        log("Downloading model from Hugging Face: \(model)")
-
-        let globs: Globs = switch backend {
-        case .llama: .init(["*\(url.lastPathComponent)"])
-        case .mlx: .mlx
-        case .foundationModels: []
-        }
-
-        let downloader = FileDownloader(source: .huggingFace(
-            id: url.pathComponents[1...2].joined(separator: "/"),
-            globs: globs
-        ))
-        try await downloader.download { progress in
-            log("Downloading model: \(progress)")
-        }
-        return switch backend {
-        case .llama: downloader.destination.appendingPathComponent(url.lastPathComponent)
-        case .mlx: downloader.destination
-        case .foundationModels: throw LocalLLMCommandError.invalidModel("Model downloading is not applicable for the FoundationModels backend.")
-        }
-        #else
-        throw LocalLLMCommandError.invalidModel("Downloading models is not supported on this platform.")
-        #endif
     }
 
     private func log(_ message: String) {
